feat: kubernetes bill of materials
